feat: FieldException | Moved related classes to separate namespace
