+++ app/src/Domain/Validators/Fields/EmailValidator.php
@@ -1,10 +1,9 @@
 <?php
 namespace Pryaniki\App\Domain\Validators\Fields;
 
-use Pryaniki\App\Exceptions;
-use Pryaniki\App\Exceptions\EmptyFieldException;
-use Pryaniki\App\Exceptions\FieldException;
-use Pryaniki\App\Exceptions\NotValidException;
+use Pryaniki\App\Exceptions\Validation\Field\EmptyFieldException;
+use Pryaniki\App\Exceptions\Validation\Field\FieldException;
+use Pryaniki\App\Exceptions\Validation\Field\NotValidException;
 use Pryaniki\App\Domain\Validators\BaseValidator;
 
 class EmailValidator extends BaseValidator
@@ -32,11 +31,11 @@ class EmailValidator extends BaseValidator
     private function checkEmail(string $email, string $fieldName): void
     {
         if ($email === '') {
-            throw new Exceptions\EmptyFieldException($fieldName);
+            throw new EmptyFieldException($fieldName);
         }
 
         if (filter_var($email, FILTER_VALIDATE_EMAIL) === false) {
-            throw new Exceptions\NotValidException($fieldName);
+            throw new NotValidException($fieldName);
         }
     }
