+++ app/src/Exceptions/Validation/Field/EmptyFieldException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pryaniki\App\Exceptions;
+namespace Pryaniki\App\Exceptions\Validation\Field;
 
 class EmptyFieldException extends FieldException
 {
