+++ app/src/Exceptions/Validation/Field/NotValidException.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace Pryaniki\App\Exceptions;
+namespace Pryaniki\App\Exceptions\Validation\Field;
+
 
 class NotValidException extends FieldException
 {
